Split isaac into 3 modules For into `isaac` and `isaac64` for easier comparison. Split of `isaac_word` to generate correct the documentation (TODO).
Nothing is changed here, just split in three files.
