+++ src/prng/mod.rs
@@ -45,8 +45,12 @@
 
 mod chacha;
 mod isaac;
+mod isaac64;
+mod isaac_word;
 mod xorshift;
 
 pub use self::chacha::ChaChaRng;
-pub use self::isaac::{IsaacRng, Isaac64Rng, IsaacWordRng};
+pub use self::isaac::IsaacRng;
+pub use self::isaac64::Isaac64Rng;
+pub use self::isaac_word::IsaacWordRng;
 pub use self::xorshift::XorShiftRng;
